2.0.3.2
Changed splitCSVLine and its related method and enum to static with the
splitCSVLine’s privacy set to package, to allow it to be unit-tested.

Wrote unit test for splitCSVLine, testing a small sample of possible
CSV lines with various peculiarities of the format.

Added assertion for the equality of lists (in addition to collections,
which were already present)

Replaced an if-else statement in assertCollectionEquals with the
equivalent use of the ternary operator, reduced the method’s method
signature declaration (along with tis overload) to a single line.