1.6.0
Events should now handle weakly registered dependent events.

Added overloads of register(…) to take a boolean to specify if whether
a dependent event should be strongly or weakly registered.

Implemented support for weakly registered dependent events in
StandardEvent, with the exception of the new register overloads, which
still have to be implemented.