4.2.0
Removed ability to create new compatibility objects from the
compatibility getter to the objects they should logically belong to.

Moved regulation to ensure that the same objects are returned for the
same calls to the platform-independent classes.

Added getters for the logical children (e.g. blocks from the world,
worlds from the server, etc.) to the platform-independent classes.

Added a static getInstance method to server, which does refer to the
compatibility getter.

Rewrote the compatibility access methods to refer to the correct
methods on objects, in the static instance of EnkiServer.

Compatibility objects are now regulated against a CacheBuilder map
rather than a MapMaker map.